fix(switch-to-chat-api): fix some errors
