--- src/main/kotlin/com/discord/bot/gptbot/representation/ChatCompletionRequest.kt
@@ -1,8 +1,0 @@
-package com.discord.bot.gptbot.representation
-
-import com.discord.bot.gptbot.representation.chat.ChatMessage
-
-data class ChatCompletionRequest(
-    val messages: List<ChatMessage>,
-    val model: String
-)
