--- src/main/kotlin/com/discord/bot/gptbot/representation/ChatCompletionResponse.kt
@@ -1,9 +1,0 @@
-package com.discord.bot.gptbot.representation
-
-import com.discord.bot.gptbot.representation.chat.ChatMessage
-
-class ChatCompletionResponse {
-    lateinit var choices: List<ChatChoice>
-
-    class ChatChoice(val message: ChatMessage)
-}
