+++ src/main/kotlin/com/discord/bot/gptbot/representation/chat/ChatCompletionRequest.kt
@@ -1,0 +1,10 @@
+package com.discord.bot.gptbot.representation.chat
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+data class ChatCompletionRequest(
+    val messages: List<ChatMessage>,
+    val model: String
+)
