+++ src/main/kotlin/com/discord/bot/gptbot/representation/chat/ChatCompletionResponse.kt
@@ -1,0 +1,9 @@
+package com.discord.bot.gptbot.representation.chat
+
+class ChatCompletionResponse {
+    lateinit var choices: List<ChatChoice>
+
+    class ChatChoice {
+        val message: ChatMessage? = null
+    }
+}
